split up database functions into separate files + other minor adjustments
