+++ cogs/__init__.py
@@ -2,3 +2,4 @@ from .archive_cog import ArchiveCog
 from .election_cog import ElectionCog
 from .events_cog import EventsCog
 from .miscellaneous_cog import MiscellaneousCog
+from .voice_cog import VoiceCog
