+++ cogs/miscellaneous_cog.py
@@ -52,4 +52,4 @@ class MiscellaneousCog(Cog, name="Miscellaneous"):
         if ctx.guild is None:
             raise Exception("ctx.guild is None")
         # if ctx.author.id != ctx.guild.owner_id:
-        await ctx.send("Boi what you tryna test 🫱")
+        await ctx.send("Boi wat you tryna test 🫱")
