+++ cogs/voice_cog.py
@@ -1,0 +1,5 @@
+from discord.ext.commands import Cog
+
+class VoiceCog(Cog, name="Voice"):
+    """ Class containing commands pertaining to the voice chat """
+    pass
