+++ database/__init__.py
@@ -1,3 +1,3 @@
-from .db_handler import log_role_change, inc_user_stat, log_server
-from .db_handler import remove_discord_server, update_next_archive_date, get_next_archive_date
-from .db_handler import update_user_stats, get_user_stats
+from .db_archive import get_next_archive_date, update_next_archive_date
+from .db_servers import log_server, remove_discord_server
+from .db_user_stats import inc_user_stat, update_user_stats, get_user_stats
